pyproject.toml
[tool.poetry]
name = "nonebot_plugin_arcaeabot"
version = "2.1.8"
description = "An arcaea plugin for nonebot2. ( A cross platform Python async bot framework. )"
authors = ["SEAFHMC <soku_ritsuki@outlook.com>"]
license = "AGPL-3.0 License"
readme = "README.md"
homepage = "https://github.com/SEAFHMC/nonebot-plugin-arcaeabot"
repository = "https://github.com/SEAFHMC/nonebot-plugin-arcaeabot"
documentation = "https://github.com/SEAFHMC/nonebot-plugin-arcaeabot/blob/master/README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = ">=2.0.0-beta.1, <=2.0.0-beta.3"
nonebot-adapter-onebot = ">=2.0.0-beta.1, <=2.1.0"
Pillow = ">=8.3.1"
peewee = ">=3.14.4"
httpx = ">=0.20.0, <1.0.0"
tqdm = "^4.1.0"
"ruamel.yaml" = "^0.17.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
